refactor: MenuList 저장위치 변경
