+++ fe/src/components/pages/Menu.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MenuList from '../\btemplate/MenuList';
+import MenuList from '../organism/MenuList';
 
 export default function Menu() {
   return (
